+++ CHANGES.md
@@ -10,7 +10,7 @@ Version 0.5 (2012-06-10)
 * improvements in C++ expansions
 * add passive popups w/ native look and feel (as for me it's better than
   custom popups in Pate, which is looks ugly w/ my color scheme)
-* use decorators to disable (show popup w/ alert) actions by document type,
+* use decorators to disable (show popup w/ alert) actions by various constraints,
   for example commenting text block w/ `#if 0' have sense only for C/C++
   documents